Task 6/requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
tqdm>=4.60.0
